# docs/SERVER_MODE/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MRSS is a modern, cross-platform desktop RSS reader with auto-translation, smart feed discovery, and AI-powered summarization.",
        "title": "MRSS API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/marcomarcogd/MRSS/issues"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "1.4.2"
    },
    "host": "localhost:1234",
    "basePath": "/api",
    "paths": {
...
        "/browser/open": {
            "get": {
                "description": "Open a URL in the user's default web browser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "browser"
                ],
                "summary": "Open URL in browser",
                "parameters": [
                    {
                        "type": "string",
                        "description": "URL to open (for GET requests)",
                        "name": "url",
                        "in": "query"
                    },
                    {
                        "description": "Open URL request (url) (for POST requests)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success status (status) or redirect URL (redirect)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (invalid URL)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Returns redirect response for client-side URL opening",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "browser"
                ],
                "summary": "Open URL in browser (server mode)",
                "parameters": [
                    {
                        "description": "Open URL request (url)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Redirect URL (redirect)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request (invalid URL)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/discovery/blogs": {
            "post": {
                "description": "Discover new blogs by analyzing friend links from a specific feed's RSS content",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "discovery"
                ],
                "summary": "Discover blogs from feed",
                "parameters": [
                    {
                        "description": "Discovery request (feed_id)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of discovered blogs",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/discovery.DiscoveredBlog"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Feed not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/discovery/single/start": {
            "post": {
                "description": "Start an asynchronous blog discovery process for a specific feed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "discovery"
                ],
                "summary": "Start single feed discovery",
                "parameters": [
                    {
                        "description": "Discovery request (feed_id)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Discovery started (status)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Feed not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Discovery already in progress",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "discovery.DiscoveredBlog": {
            "type": "object",
            "properties": {
                "homepage": {
                    "type": "string"
                },
                "icon_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "recent_articles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/discovery.RecentArticle"
                    }
                },
                "rss_feed": {
                    "type": "string"
                }
            }
        },
...
        "discovery.RecentArticle": {
            "type": "object",
            "properties": {
                "date": {
                    "description": "ISO 8601 format or relative time",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}